fix files ignore mechanism.

+++ src/Commands/GenerateClient.php
@@ -150,35 +150,36 @@ abstract class GenerateClient extends Command
         return $resultPath;
     }
 
-    private function recursiveClearDirectory(string $dir, int $level = 0)
+    private function recursiveClearDirectory(string $dir, int $level = 0, string $baseDir = ''): bool
     {
         if (!$dir) {
-            return;
+            return true;
         }
 
-        if ($level === 0 && !is_dir($dir)) {
-            mkdir($dir, 0755, true);
-        }
-
+        $disableDeleteDir = false;
         foreach (scandir($dir) as $fileWithoutDir) {
             if (in_array($fileWithoutDir, ['..', '.'])) {
                 continue;
             }
             $file = $dir . "/" . $fileWithoutDir;
+            $pathFromBase = $baseDir ? $baseDir . '/' . $fileWithoutDir : $fileWithoutDir;
 
-            if ($level === 0 && in_array($fileWithoutDir, $this->filesToIgnoreDuringCleanup)) {
+            if (in_array($pathFromBase, $this->filesToIgnoreDuringCleanup)) {
+                $disableDeleteDir = true;
                 continue;
             }
 
             if (is_dir($file)) {
-                $this->recursiveClearDirectory($file, $level + 1);
+                $disableDeleteDir = $this->recursiveClearDirectory($file, $level + 1, $pathFromBase) || $disableDeleteDir;
             } else {
                 unlink($file);
             }
         }
 
-        if ($level > 0) {
+        if ($level > 0 && !$disableDeleteDir) {
             rmdir($dir);
         }
+
+        return $disableDeleteDir;
     }
 }
